requirements.txt
ollamas
psycopg2
python-dotenv
streamlit
pypdf2